% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_system}
\alias{simulate_system}
\title{simulate_system}
\usage{
simulate_system(compartments, q, ec50, n, kin, diff_coeff, tau, omega, t, inc)
}
\arguments{
\item{compartments}{Tow-dimensional representation of model compartments.}

\item{q}{Clearance of the capsaicin conentration in the peripheral compartment of the subcutaneous fat tissue.}

\item{ec50}{Effect activity triggered by capsaicin concentration of the peripheral compartment in the transit compartment.}

\item{n}{The Hill coefficient as a measure of ultrasensitivity.}

\item{kin}{Apparent effect production rate in the transit compartment.}

\item{diff_coeff}{Lateral diffusion coefficient for the pain stimulus diffusion within the transit compartment.}

\item{tau}{Decay time that describes the decay of the effect in the transit compartment.}

\item{omega}{Threshold of severity perception of a pain stimulus.}

\item{t}{Duration of the simulation.}

\item{inc}{Number of simulated increments.}
}
\value{
Tow-dimensional representation of model compartments. 4 Compartments.
}
\description{
Simulate the time evaluation of the diffusion of intracutaneously applied capsaicin-evoked pain mediated by the spinal cord.
}
\details{
The model is build from 4 compartments. 1 The peripheral compartment; 2 The central compartment; 3 The transit compartment; 4 The effect compartment
}
